content updates & pdf formatting tweaks
- made the parentsquare section less anemic (hopefully not too much
  less)
- increased the font size on the pdf to 14 and played with the margins
  to get the sections to roughly align to pages (padded the right margin
  more than the left to get it to feel balanced)
- little wording tweaks, it's perfect now

note: generating the pdf automatically on commit is great and
theoretically what we want but in practice i want to generate it
manually and tweak the parameters (eg margins) to make sure it looks
nice. probably we want a package.json with the md-to-pdf command so we
don't have to dig it out of .git/hooks/pre-commit.