split to cv and resume
i got feedback that a cv is not a resume and actually the latter is the
important one for getting a job (basically, keep it short).  so this
commit moves the old content to `cv.md` and adds a resume and a separate
top-level readme.  it also moves the md-to-pdf config into frontmatter
on the two files (since they have different margins/css/whatever),
removes the old config file, and updates the scripts in package.json.